WIP: Initial Whonix 11 + jessie commit

+++ whonix-gateway/02_install_groups_pre.sh
@@ -1,6 +1,40 @@
 #!/bin/bash -e
 # vim: set ts=4 sw=4 sts=4 et :
 
+# ==============================================================================
+#                           WHONIX 11 - WIP NOTES
+# ==============================================================================
+# 
+# TODO - TEMP TEST:
+# ------------------------------------------------------------------------------
+# -
+#
+# TODO - EXPERIMENT:
+# ------------------------------------------------------------------------------
+# -
+#
+# TODO - FIX:
+# ------------------------------------------------------------------------------
+# -
+#
+# BUGS:
+# ------------------------------------------------------------------------------
+# 1700_install-packages:450 --> "$WHONIX_SOURCE_HELP_STEPS_FOLDER/remove-local-temp-apt-repo"
+# INFO: Setting... export UWT_DEV_PASSTHROUGH="1"
+# INFO: Variable anon_dist_build_version was already set to: 11.0.0.0.1
+# /home/user/Whonix/help-steps/pre: line 20: error_: command not found
+# ...
+# + true 'INFO: Currently running script: /home/user/Whonix/help-steps/unprevent-daemons-from-starting '
+# + true 'INFO: Currently running script: /home/user/Whonix/help-steps/unchroot-raw '
+# + true 'INFO: Skipping script, because ANON_BUILD_INSTALL_TO_ROOT=1: /home/user/Whonix/help-steps/unmount-raw'
+# + true 'INFO: Currently running script: ././build-steps.d/2300_run-chroot-scripts-post-d '
+#
+# REPAIR AFTER SUCCESSFUL BUILDS:
+# ------------------------------------------------------------------------------
+# -
+# 
+# ==============================================================================
+
 source "${SCRIPTSDIR}/vars.sh"
 source "${SCRIPTSDIR}/distribution.sh"
 
@@ -17,59 +51,119 @@ trap cleanup EXIT
 
 if ! [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_prepared_groups" ]; then
     #### '----------------------------------------------------------------------
-    info ' Installing extra packages in packages_whonix.list file'
+    info ' Installing extra packages from Whonix 30_dependencies'
     #### '----------------------------------------------------------------------
-    installPackages packages_whonix.list
+    source "${WHONIX_DIR}/buildconfig.d/30_dependencies"
+    aptInstall ${whonix_build_script_build_dependency}
+
     touch "${INSTALLDIR}/${TMPDIR}/.whonix_prepared_groups"
 fi
 
 
-# ------------------------------------------------------------------------------
+# ==============================================================================
 # chroot Whonix build script
-# ------------------------------------------------------------------------------
+# ==============================================================================
 read -r -d '' WHONIX_BUILD_SCRIPT <<'EOF' || true
 ################################################################################
-# Pre Fixups
-sudo mkdir -p /boot/grub2
-sudo touch /boot/grub2/grub.cfg
-sudo mkdir -p /boot/grub
-sudo touch /boot/grub/grub.cfg
-sudo mkdir --parents --mode=g+rw "/tmp/uwt"
+# This script is executed from chroot most likely as the user `user`
+# 
+# - The purpose is to do a few pre-fixups that are directly related to whonix
+#   build process
+# - Then, finally, call `whonix_build_post` as sudo with a clean (no) ENV
+#
+################################################################################
 
-# Whonix seems to re-install sysvinit even though there is a hold
-# on the package.  Things seem to work anyway. BUT hopfully the
-# hold on grub* don't get removed
-sudo apt-mark hold sysvinit
-sudo apt-mark hold grub-pc grub-pc-bin grub-common grub2-common
+
+# ==============================================================================
+# TODO: Confirm if these fixups still apply to Whonix 11
+# ==============================================================================
+
 
 # Whonix expects haveged to be started
+# XXX: What to do when init file gone since can not start systed in chroot ENV?
+# ------------------------------------------------------------------------------
 sudo /etc/init.d/haveged start
-################################################################################
-# Whonix installation
-export WHONIX_BUILD_UNATTENDED_PKG_INSTALL="1"
 
-pushd ~/Whonix
-sudo ~/Whonix/whonix_build \
+# Use sudo with clean ENV to build Whonix; any ENV options will be set there
+# ------------------------------------------------------------------------------
+sudo ~/whonix_build_post $@
+EOF
+
+# ==============================================================================
+# chroot Whonix post build script
+# ==============================================================================
+read -r -d '' WHONIX_BUILD_SCRIPT_POST <<'EOF' || true
+#!/bin/bash -e
+# vim: set ts=4 sw=4 sts=4 et :
+
+
+# =============================================================================
+# `TO KEEP` COMMENTED OUT CONFIGURATIONS
+# =============================================================================
+# Make sure we clear Qubes overrides of these vars
+# - These will be kept for future reference in case module is re-factored
+#   differently and would be required if ENV was passed via sudo
+#export GENMKFILE_INCLUDE_FILE_MAIN=
+#export GENMKFILE_BOOTSTRAP=
+
+
+# =============================================================================
+# `TO REMOVE` CONFIGURATIONS + HACK THAT WILL MOST LIKELY BE REMOVED
+# =============================================================================
+# Whonix 11 Hacks (stretch does not exist)
+# XXX: Should be fixed in next tag release
+# ------------------------------------------------------------------------------
+export whonix_build_apt_newer_release_codename="jessie"
+
+# Disable lintian; cause too many build errors
+# XXX: Will be fixed at some point when lintian error have been fixed for jessie
+# ------------------------------------------------------------------------------
+# `sed` only needed till next tag release
+sudo sed -i "s/make_use_lintian=\"true\"/make_use_lintian=\"false\"/g" "/home/user/Whonix/build-steps.d/1200_create-debian-packages"
+export make_use_lintian="true"
+
+
+# =============================================================================
+# `REQUIRED` CONFIGURATIONS
+# =============================================================================
+# Prevents Whonix makefile use of shared memory 'sem_open: Permission denied'
+# ------------------------------------------------------------------------------
+echo tmpfs /dev/shm tmpfs defaults 0 0 >> /etc/fstab
+mount /dev/shm
+
+
+# =============================================================================
+# WHONIX BUILD COMMAND
+# =============================================================================
+pushd /home/user/Whonix
+/home/user/Whonix/whonix_build \
     --flavor $1 \
     -- \
     --build \
     --arch amd64 \
-    --kernel linux-image-amd64 \
-    --headers linux-headers-amd64 \
-    --target root \
     --freshness current \
+    --target root \
     --report minimal \
     --verifiable minimal \
+    --allow-uncommitted true \
+    --allow-untagged true \
     --sanity-tests false || { exit 1; }
 popd
 EOF
 
+
+# Whonix11 removed for now...
+# ====================================
+#    --kernel linux-image-amd64 \
+#    --headers linux-headers-amd64 \
+#
 # Some Additional Whonix build options
 # ====================================
-# --tb close  # Install tor-browser
-# --allow-uncommitted true
-# --allow-untagged true
-# --testing-frozen-sources  # Jessie; no current sources
+#    --tb close  # Install tor-browser \
+#    --allow-uncommitted true \
+#    --allow-untagged true \
+#    --testing-frozen-sources  # Jessie; no current sources \
+
 
 ##### '-------------------------------------------------------------------------
 debug ' Preparing Whonix for installation'
@@ -85,14 +179,6 @@ if [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_prepared_groups" ] && ! [ -f "${INSTALL
         su $(logname) -c "git submodule update --init --recursive";
     }
     popd
-
-    #### '----------------------------------------------------------------------
-    info ' Faking grub installation since Whonix has depends on grub-pc'
-    #### '----------------------------------------------------------------------
-    mkdir -p "${INSTALLDIR}/boot/grub"
-    cp "${INSTALLDIR}/usr/lib/grub/i386-pc/"* "${INSTALLDIR}/boot/grub"
-    rm -f "${INSTALLDIR}/usr/sbin/update-grub"
-    chroot ln -fs /bin/true /usr/sbin/update-grub
 
     #### '----------------------------------------------------------------------
     info ' Adding a user account for Whonix to build with'
@@ -111,14 +197,24 @@ if [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_prepared_groups" ] && ! [ -f "${INSTALL
     #### '----------------------------------------------------------------------
     info ' Installing Whonix build scripts'
     #### '----------------------------------------------------------------------
-    echo "${WHONIX_BUILD_SCRIPT}" > "${INSTALLDIR}/home/user/whonix_build.sh"
-    chmod 0755 "${INSTALLDIR}/home/user/whonix_build.sh"
+    echo "${WHONIX_BUILD_SCRIPT_POST}" > "${INSTALLDIR}/home/user/whonix_build_post"
+    chmod 0755 "${INSTALLDIR}/home/user/whonix_build_post"
+
+    echo "${WHONIX_BUILD_SCRIPT}" > "${INSTALLDIR}/home/user/whonix_build"
+    chmod 0755 "${INSTALLDIR}/home/user/whonix_build"
 
     #### '----------------------------------------------------------------------
     info ' Removing apt-listchanges if it exists,so no prompts appear'
     #### '----------------------------------------------------------------------
     #      Whonix does not handle this properly, but aptInstall packages will
     aptRemove apt-listchanges || true
+
+    #### '----------------------------------------------------------------------
+    debug 'XXX: Whonix11 HACK since we running all sections without conditions'
+    debug '     and we are deleting and linking interfaces below'
+    debug 'XXX: Remove interfaces from files directory when this hack not needed'
+    #### '----------------------------------------------------------------------
+    rm -f "${INSTALLDIR}/etc/network/interfaces"
 
     #### '----------------------------------------------------------------------
     info ' Copying additional files required for build'
@@ -137,6 +233,7 @@ if [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_prepared" ] && ! [ -f "${INSTALLDIR}/${
         chroot su user -c 'mkdir /home/user/Whonix'
     fi
 
+    # XXX: TODO: Need to get from ENV $WHONIX_DIR as we can't always be sure where dir is
     mount --bind "../Whonix" "${INSTALLDIR}/home/user/Whonix"
 
     if [ "${TEMPLATE_FLAVOR}" == "whonix-gateway" ]; then
@@ -154,7 +251,15 @@ if [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_prepared" ] && ! [ -f "${INSTALLDIR}/${
     mount --bind /dev "${INSTALLDIR}/dev"
     mount --bind /dev/pts "${INSTALLDIR}/dev/pts"
 
-    chroot su user -c "cd ~; ./whonix_build.sh ${BUILD_TYPE} ${DIST}" || { exit 1; }
+    info 'Executing whonix_build script now...'
+    chroot su user -c "cd ~; ./whonix_build ${BUILD_TYPE} ${DIST}" || { exit 1; }
+
+#    # Issues with logger; revert back to no logging for now
+#    warn 'Executing whonix_build script now...'
+#    # XXX: Remove static reference and store in template-whonix $DIR
+#    BUILD_LOG=/home/user/qubes/qubes-src/template-whonix/whonix.log
+#    chroot su user -c "cd ~; ./whonix_build ${BUILD_TYPE} ${DIST}" 3>&2 2>&1 | tee -a ${BUILD_LOG} || { exit 1; }
+
     touch "${INSTALLDIR}/${TMPDIR}/.whonix_installed"
 fi
 
@@ -169,19 +274,23 @@ if [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_installed" ] && ! [ -f "${INSTALLDIR}/$
     #### '----------------------------------------------------------------------
     pushd "${INSTALLDIR}/etc/network"
     {
-        rm -f interfaces;
-        ln -s interfaces.backup interfaces;
+        if [ -e 'interfaces.backup' ]; then
+            rm -f interfaces;
+            ln -s interfaces.backup interfaces;
+        fi
     }
     popd
 
     #### '----------------------------------------------------------------------
-    info ' Temporarily retore original resolv.conf for remainder of install process'
-    info ' (Will be restored back in wheezy+whonix/04_qubes_install_post.sh)'
+    info ' Temporarily restore original resolv.conf for remainder of install process'
+    info ' (Will be restored back in jessie+whonix/04_qubes_install_post.sh)'
     #### '----------------------------------------------------------------------
     pushd "${INSTALLDIR}/etc"
     {
-        rm -f resolv.conf;
-        cp -p resolv.conf.backup resolv.conf;
+        if [ -e 'resolv.conf.backup' ]; then
+            rm -f resolv.conf;
+            cp -p resolv.conf.backup resolv.conf;
+        fi
     }
     popd
 
@@ -191,8 +300,10 @@ if [ -f "${INSTALLDIR}/${TMPDIR}/.whonix_installed" ] && ! [ -f "${INSTALLDIR}/$
     #### '----------------------------------------------------------------------
     pushd "${INSTALLDIR}/etc"
     {
-        rm -f hosts;
-        cp -p hosts.anondist-orig hosts;
+        if [ -e 'hosts.anondist-orig' ]; then
+            rm -f hosts;
+            cp -p hosts.anondist-orig hosts;
+        fi
     }
     popd
